Revert "fix inventory, energy link"
This reverts commit 1e5bdccf65161c59990dec52fb4705d5baebfead.

+++ Assets/script/Building/BuildingReceivedEnergy.cs
@@ -7,7 +7,7 @@ public class BuildingReceivedEnergy : MonoBehaviour, ISubscrireEvent
 {
     public List<GenerateEnergy> ConnectGenerator = new();
     public bool IsLinked;
-    [SerializeField, ReadOnly(true)] private float _sumOfEnergyReceived;
+    [SerializeField, ReadOnly(true)] private int _sumOfEnergyReceived;
 
     private BuildingManager instance;
     public void SubscrireEvent()
